backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
pymongo==4.4.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
email-validator==2.0.0
requests==2.31.0
